pyproject.toml
[tool.poetry]
name = "italki-teachers"
version = "0.1.0"
description = ""
authors = ["Luke"]
readme = "README.md"
packages = [{include = "italki_teachers"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
python-dateutil = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
